text adventure FINISHED
Not sure about the streamline printing requirement, as that was already something implemented in enums v2, but I think I got everything to get me to a B. I tried to implement a locked door and an npc, but I just had no idea how to do it. I ended up having it be a code that lets you in a room, but the code was a command, so if you type help, or in my case cheat, it shows you the code, which takes away the purpose of exploring the rest of the map. The CommandWords class was quite confusing for me to touch around, so I hope I didnt break anything there. Essentially, it is super easy to escape my space prison that you would think these aliens are eithier really stupid or just don't care. All in all, I hope it is satisfactory.